beginning refactor
