+++ Cargo.toml
@@ -10,6 +10,7 @@ edition = "2018"
 crate-type = ["cdylib"]
 
 [dependencies]
+lazy_static = "1.4.0"
 mlua = {version = "0.6", features = ["luajit", "vendored", "module", "macros", "send", "async"]}
 roxmltree = "0.14.1"
 regex = "1"
