+++ src_old/constants.rs
@@ -1,0 +1,3 @@
+pub const DASH_APP_BASE_PATH: &str = "/Applications/Dash.app";
+pub const DASH_APP_CLI_PATH: &str = "/Contents/Resources/dashAlfredWorkflow";
+pub const KEYWORD_PATTERN: &str = r"^([a-zA-Z]+):.+";
